+++ backend/src/main/java/fr/projet/diginamic/backend/services/interfaces/ServiceInterface.java
@@ -5,11 +5,11 @@ import java.util.List;
 public interface ServiceInterface<T, DTO> {
     public List<T> getAll();
 
-    public T getOne(Integer id);
+    public T getOne(Long id);
 
     public List<T> create(DTO userDto);
 
-    public List<T> update(Integer id, DTO userDto);
+    public List<T> update(Long id, DTO userDto);
 
-    public List<T> delete(Integer id);
+    public List<T> delete(Long id);
 }
